Move WindowBase to the Indicators namespace
